Pom.xml is organized better. =)
