+++ src/main/java/com/travelstory/dao/UserDAO.java
@@ -8,7 +8,6 @@ import org.springframework.stereotype.Repository;
 public interface UserDAO extends JpaRepository<User, Long> {
     boolean existsByEmail(String email);
 
-
     boolean existsByEmailAndPassword(String email, String password);
 
     User findByEmail(String email);
