+++ src/main/java/com/travelstory/services/UserServiceImpl.java
@@ -6,7 +6,7 @@ import com.travelstory.dto.RegistrationDTO;
 import com.travelstory.entity.User;
 import com.travelstory.entity.UserRole;
 import com.travelstory.exceptions.ResourceNotFoundException;
-import com.travelstory.sequrity.TokenProvider;
+import com.travelstory.security.TokenProvider;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -37,7 +37,8 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User updateAvatar(Long userId, User userDetails) {
-        User user = userDAO.findById(userId).orElseThrow(() -> new ResourceNotFoundException("User", "id", userId));
+        User user = userDAO.findById(userId).orElseThrow(() ->
+                new ResourceNotFoundException("User", "id", userId));
 
         user.setAvatar(userDetails.getAvatar());
